+++ foobar2000/foo_uie_wsh_panel_mod/foo_uie_wsh_panel_mod/script_interface_impl.cpp
@@ -1334,7 +1334,7 @@ STDMETHODIMP FbMetadbHandle::get_SubSong(UINT* p)
 	return S_OK;
 }
 
-STDMETHODIMP FbMetadbHandle::get_FileSize(LONGLONG* p)
+STDMETHODIMP FbMetadbHandle::get_FileSize(double* p)
 {
 	TRACK_FUNCTION();
